requirements.txt
absl-py==0.9.0
affine==2.3.0
aiohttp==3.6.2
alibi==0.5.0
anatome @ git+https://github.com/moskomule/anatome@1f3e602d5e5dc9656c76cdf59338bdb744013f74
appdirs==1.4.4
appnope==0.1.0
arrayfire==3.6.20181017
astor==0.8.1
async-timeout==3.0.1
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.9.0
bleach==3.1.0
blis==0.4.1
cached-property==1.5.2
cachetools==4.1.0
-e git+https://github.com/pytorch/captum.git@794b55302f16e1e9ed2f10379a6989bc7a5378b4#egg=captum
catalogue==1.0.0
certifi==2019.9.11
chardet==3.0.4
click==7.1.1
click-plugins==1.1.1
cligj==0.7.1
cloudpickle==1.3.0
cocos==0.2.2
colorama==0.4.3
colorful==0.5.4
contexttimer==0.3.3
cubature==0.14.5
cycler==0.10.0
cymem==2.0.3
Cython==0.29.21
dcor==0.5.3
decorator==4.4.0
defusedxml==0.6.0
dictances==1.5.3
dill==0.3.3
divergence==0.4.2
eagerpy==0.27.0
entrypoints==0.3
filelock==3.0.12
future==0.18.2
gabornet==0.2.0
gast==0.2.2
gitdb==4.0.2
GitPython==3.1.0
google==2.0.3
google-auth==1.14.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
graphviz==0.16
grpcio==1.28.1
h5py==2.10.0
idna==2.8
idna-ssl==1.1.0
image-similarity-measures==0.3.4
imageio==2.6.1
importlib-metadata==1.7.0
iniconfig==1.1.1
ipykernel==5.1.3
ipython==7.9.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.15.1
Jinja2==2.10.3
joblib==0.14.0
jsonpatch==1.24
jsonpointer==2.0
jsonschema==3.1.1
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lime==0.2.0.1
llvmlite==0.35.0
loky==2.9.0
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.3.4
matplotlib-scalebar==0.7.2
mistune==0.8.4
mlxtend==0.17.3
mnist==0.2.2
more-itertools==7.2.0
mpmath==1.1.0
multidict==4.7.5
multiprocess==0.70.11.1
murmurhash==1.0.2
nbconvert==5.6.1
nbformat==4.4.0
nbstripout==0.3.6
networkx==2.4
nose==1.3.7
notebook==6.0.1
numba==0.52.0
numexpr==2.7.1
numpy==1.21.1
oauthlib==3.1.0
onnx==1.6.0
-e git+git@github.com:onnx/onnx-tensorflow.git@c4fc047e7ec71daa6aa8f71e9cc2ee9e5a3768b6#egg=onnx_tf
onnxruntime==1.2.0
opencv-python==4.3.0.36
opt-einsum==3.2.1
packaging==20.9
palettable==3.3.0
pandas==1.2.1
pandocfilters==1.4.2
parso==0.5.1
pathos==0.2.7
patsy==0.5.1
pexpect==4.7.0
phasepack==1.5
pickleshare==0.7.5
Pillow==6.2.1
plac==1.1.3
pluggy==0.13.1
pooch==1.3.0
POT==0.7.0
pox==0.2.9
ppft==1.6.6.3
preshed==3.0.2
prettyprinter==0.18.0
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.11.3
ptyprocess==0.6.0
py==1.10.0
py-spy==0.3.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pydicom==1.4.0
pydot==1.4.1
pyFFTW==0.12.0
Pygments==2.4.2
pygraphviz==1.5
pyparsing==2.4.2
pyrsistent==0.15.5
pytest==6.2.2
python-dateutil==2.8.0
python-dotenv==0.15.0
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==18.1.0
qtconsole==4.5.5
rasterio==1.1.8
ray==0.8.4
redis==3.4.1
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-image==0.17.2
scikit-learn==0.24.2
scikit-posthocs==0.6.7
scipy==1.7.0
seaborn==0.11.1
seaborn-image==0.4.2
secml==0.11.2
Send2Trash==1.5.0
shap==0.35.0
six==1.12.0
sklearn==0.0
skorch==0.9.0
smmap==3.0.1
snuggs==1.4.7
soupsieve==2.0
spacy==2.3.1
srsly==1.0.2
statsmodels==0.12.1
sympy==1.4
tables==3.6.1
tabulate==0.8.7
tensorboard==2.4.1
tensorboard-plugin-wit==1.6.0.post3
tensorboardX==2.0
tensorflow==1.15.3
tensorflow-addons==0.9.1
tensorflow-estimator==1.15.1
tensorflow-probability==0.9.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
thinc==7.4.1
threadpoolctl==2.1.0
tifffile==2021.2.26
toml==0.10.2
torch==1.6.0
torchcontrib==0.0.2
torchfile==0.1.0
torchvision==0.7.0
torchviz==0.0.2
tornado==6.0.3
tqdm==4.36.1
traitlets==4.3.3
typeguard==2.7.1
typing-extensions==3.7.4.2
universal-divergence==0.2.0
urllib3==1.25.6
visdom==0.1.8.9
Wand==0.6.7
wasabi==0.7.0
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.56.0
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
yarl==1.4.2
zipp==0.6.0
